Split report summary presentation helpers
