+++ apps/server/vibesensor/use_cases/history/report_document/evidence_snapshot.py
@@ -8,8 +8,8 @@ from typing import TYPE_CHECKING
 from vibesensor.domain.diagnosis_assessment import LEGACY_CONTEXT_CAVEAT_KEY
 from vibesensor.shared.boundaries.reporting import PreparedReportFacts
 from vibesensor.shared.boundaries.reporting.document import ReportLabelValueRow
+from vibesensor.shared.report_confidence_presentation import confidence_snapshot_text
 from vibesensor.shared.report_presentation import (
-    confidence_snapshot_text,
     display_location,
     human_source,
 )
